fix(model): fix thread max message count (#1822)
Change the type of channel's message_count to u64 to allow for larger
integers, since Discord is starting to track more detailed counts. I've
recently been receiving integers much larger than 50, such that a u8
would not suffice. This is currently not documented, however, I have
clarified with a Discord staff and it is the intended behaviour.